refactored Model directory
