+++ RamenShopApp/Model/User/Profile.swift
@@ -1,0 +1,15 @@
+//
+//  Profile.swift
+//  RamenShopApp
+//
+//  Created by Koro Saka on 2021-02-20.
+//  Copyright © 2021 Koro Saka. All rights reserved.
+//
+
+import Foundation
+import Firebase
+
+struct Profile {
+    var userName: String
+    var icon: UIImage?
+}
